// src/app/admin/system/goals/goals.interface.ts
export interface IGoalMonth {
  id:           number;
  yearmonth:    number;
  year:         number,
  month:        number;
  total:        number;
  name:         string;
}

export interface IGoalWeek {
  id:           number;
  yearweek:     number;
  year:         number,
  week:         number;
  total:        number;
}